split Pedersen commitments out into their own module
