+++ .gitignore
@@ -6,6 +6,8 @@ bench_verify
 bench_schnorr_verify
 bench_recover
 bench_internal
+bench_generator
+bench_rangeproof
 tests
 exhaustive_tests
 gen_context
