+++ Makefile.am
@@ -186,6 +186,10 @@ if ENABLE_MODULE_GENERATOR
 include src/modules/generator/Makefile.am.include
 endif
 
+if ENABLE_MODULE_COMMITMENT
+include src/modules/commitment/Makefile.am.include
+endif
+
 if ENABLE_MODULE_RANGEPROOF
 include src/modules/rangeproof/Makefile.am.include
 endif
