+++ src/modules/commitment/pedersen_impl.h
@@ -4,8 +4,8 @@
  * file COPYING or http://www.opensource.org/licenses/mit-license.php. *
  ***********************************************************************/
 
-#ifndef _SECP256K1_PEDERSEN_IMPL_H_
-#define _SECP256K1_PEDERSEN_IMPL_H_
+#ifndef SECP256K1_MODULE_COMMITMENT_PEDERSEN
+#define SECP256K1_MODULE_COMMITMENT_PEDERSEN
 
 #include <string.h>
