+++ src/secp256k1.c
@@ -23,10 +23,12 @@
 # include "include/secp256k1_generator.h"
 #endif
 
+#ifdef ENABLE_MODULE_COMMITMENT
+# include "include/secp256k1_commitment.h"
+#endif
+
 #ifdef ENABLE_MODULE_RANGEPROOF
 # include "include/secp256k1_rangeproof.h"
-# include "modules/rangeproof/pedersen.h"
-# include "modules/rangeproof/rangeproof.h"
 #endif
 
 #define ARG_CHECK(cond) do { \
@@ -608,6 +610,10 @@ int secp256k1_ec_pubkey_combine(const secp256k1_context* ctx, secp256k1_pubkey *
 
 #ifdef ENABLE_MODULE_GENERATOR
 # include "modules/generator/main_impl.h"
+#endif
+
+#ifdef ENABLE_MODULE_COMMITMENT
+# include "modules/commitment/main_impl.h"
 #endif
 
 #ifdef ENABLE_MODULE_RANGEPROOF
